Added Orientation.h and Priority.h Moved orientation and direction to Orientation.h Moved priority to Priority.h Removed struct_direction_list as it was a duplicate of build_orientations
